feat: Java osv-scanner Maven dep-vuln + SSOT extraction (10k.1.4)
Capabilities pipeline complete (5/5). Mirrors the 10k.1.2 JaCoCo
SSOT extraction pattern: kotlin's parser was already
ecosystem-agnostic (Maven-filtered, not Kotlin-coupled), so the
refactor is "relocate to shared module + delegate from both packs",
not "fork and modify".

Shared module: `src/analyzers/tools/osv-scanner-maven.ts` —
- `parseOsvScannerMavenFindings(raw)` — pure JSON parser, ecosystem-
  filtered to Maven so polyglot repos don't double-count between
  packs (typescript pack still owns npm, python pack owns PyPI, etc.)
- `gatherOsvScannerMavenDepVulnsResult(cwd, source)` — manifest
  discovery (gradle.lockfile > pom.xml > verification-metadata) +
  tool invocation + CVSS alias-fallback resolution. `source` arg
  reserved for future per-pack provenance attribution.

Kotlin pack: dropped 170+ lines of inline depVulns logic.
`kotlinDepVulnsProvider.gather` is a 1-line delegate. Test imports
of `parseOsvScannerMavenFindings` rebased to the new module.

Java pack: `javaDepVulnsProvider` registered. tools[] now declares
`['pmd', 'osv-scanner']`. cliBinaries unchanged (java + mvn + pmd).

Real fixture: `test/fixtures/benchmarks/java/pom.xml` with
- commons-collections:3.2.1 (CVE-2015-7501 — the original "Mad
  Gadget" Java deserialization exploit)
- log4j-core:2.14.0 (CVE-2021-44228 Log4Shell — already used by
  kotlin fixture; keeping it here for cross-pack consistency)

Different from kotlin's gson-only choice deliberately — distinct
deps make polyglot regressions more meaningful when 10i fingerprint
work lands.

Captured `test/fixtures/raw/java/osv-scanner-output.json` (9
findings: 1 commons-collections + 8 log4j-core advisories).

Cross-ecosystem matrix: new
`describe('cross-ecosystem benchmarks — Java')` mirroring kotlin's,
asserts the commons-collections finding by package + version
(advisory id may rotate but package@version is stable).

End-to-end smoke: `dxkit vulnerabilities test/fixtures/benchmarks/java`
surfaces 9 dependency findings (3 critical, 2 high, 4 medium) with
full enrichment (CVSS, EPSS, references).

Recipe v3 reflection: BEST CASE SSOT pattern — same as 10k.1.2.
Parser was already generic from day 1 (kotlin's tests pre-validated
this). Refactor was relocation, not modification. No new G-class
gaps surfaced. The sibling-pack-overlap-on-build-manifest concern
from G9 didn't repeat here because detect was already source-driven
after 10k.1.3's fix.

Tests: 957 passed | 0 skipped (52/52 files). Capabilities pipeline:
**5 of 5 done**. Next: 10k.1.5 — version bump + 2.4.5 ship.
